requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
playwright==1.40.0
pandas==2.1.3
pdfplumber==0.10.3
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
openpyxl==3.1.2
xlrd==2.0.1

